+++ app/server/controllers/VideoController.js
@@ -323,7 +323,7 @@ VideoController.prototype.addEntityAction = function(request, response, next) {
             var validationDescriptor = {
               title: {type: 'string', required: true},
               date: {type: 'number', default: Date.now()},
-              shortDescription: {type: 'string'},
+              leadParagraph: {type: 'string'},
               description: {type: 'string'},
               groups: {type: 'array<string>', in: groupsIds}
             };
@@ -375,7 +375,7 @@ VideoController.prototype.addEntityAction = function(request, response, next) {
             originalFileName: pathDescriptor.name,
             title: params.title,
             date: params.date,
-            shortDescription: params.shortDescription,
+            leadParagraph: params.leadParagraph,
             description: params.description,
             category: params.category,
             groups: params.groups,